Updated code towards roam implementation.
* added more roam code
* updated mesh code
* updated vector code

+++ Simulator/Simulator/Simulator.Shared/Base/Math/Vector.cpp
@@ -3,16 +3,16 @@
 #include "Vector.h"
 
 
-Base::Math::Vector Base::Math::VectorInit(std::initializer_list<double> l)
+Base::Math::Vector Base::Math::VectorInit(std::initializer_list<float> l)
 {
-	double tmp[4] = { 0, 0, 0, 0 };
+	float tmp[4] = { 0, 0, 0, 0 };
 	if (l.size() <= 4)
 	{
 		int n = 0;
-		for (const double* it = begin(l); it != end(l); ++it, ++n)
+		for (const float* it = begin(l); it != end(l); ++it, ++n)
 		{
 			tmp[n] = *it;
 		}
 	}
-	return DirectX::XMVectorSet((float)tmp[0], (float)tmp[1], (float)tmp[2], (float)tmp[3]);
+	return DirectX::XMVectorSet(tmp[0], tmp[1], tmp[2], tmp[3]);
 }
